requirements.txt
langchain
tiktoken
lark
hydra-core
python-dotenv
tqdm
chromadb
openai
